+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @debbl/eslint-config-basic
 
+## 0.0.28
+
+### Patch Changes
+
+- prettier printWidth
+
 ## 0.0.27
 
 ### Patch Changes