+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @debbl/eslint-config-react
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @debbl/eslint-config-ts@0.0.28
+
 ## 0.0.27
 
 ### Patch Changes